added support for TCP Socket idle keepalives to client
(server is next)

+++ svr-auth.c
@@ -1,19 +1,19 @@
 /*
  * Dropbear - a SSH2 server
- * 
+ *
  * Copyright (c) 2002,2003 Matt Johnston
  * All rights reserved.
- * 
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -103,7 +103,7 @@ void recv_msg_userauth_request() {
 	if (servicelen != SSH_SERVICE_CONNECTION_LEN
 			&& (strncmp(servicename, SSH_SERVICE_CONNECTION,
 					SSH_SERVICE_CONNECTION_LEN) != 0)) {
-		
+
 		/* TODO - disconnect here */
 		m_free(username);
 		m_free(servicename);
@@ -111,7 +111,7 @@ void recv_msg_userauth_request() {
 		dropbear_exit("unknown service in auth");
 	}
 
-	/* check username is good before continuing. 
+	/* check username is good before continuing.
 	 * the 'incrfail' varies depending on the auth method to
 	 * avoid giving away which users exist on the system through
 	 * the time delay. */
@@ -127,10 +127,10 @@ void recv_msg_userauth_request() {
 		if (valid_user
 				&& svr_opts.allowblankpass
 				&& !svr_opts.noauthpass
-				&& !(svr_opts.norootpass && ses.authstate.pw_uid == 0) 
-				&& ses.authstate.pw_passwd[0] == '\0') 
+				&& !(svr_opts.norootpass && ses.authstate.pw_uid == 0)
+				&& ses.authstate.pw_passwd[0] == '\0')
 		{
-			dropbear_log(LOG_NOTICE, 
+			dropbear_log(LOG_NOTICE,
 					"Auth succeeded with blank password for '%s' from %s",
 					ses.authstate.pw_name,
 					svr_ses.addrstring);
@@ -144,7 +144,7 @@ void recv_msg_userauth_request() {
 			goto out;
 		}
 	}
-	
+
 #if DROPBEAR_SVR_PASSWORD_AUTH
 	if (!svr_opts.noauthpass &&
 			!(svr_opts.norootpass && ses.authstate.pw_uid == 0) ) {
@@ -335,7 +335,7 @@ static int checkusername(const char *username, unsigned int userlen) {
 	dropbear_log(LOG_WARNING, "User '%s' has invalid shell, rejected",
 				ses.authstate.pw_name);
 	return DROPBEAR_FAILURE;
-	
+
 goodshell:
 	endusershell();
 	TRACE(("matching shell"))
@@ -357,7 +357,7 @@ void send_msg_userauth_failure(int partial, int incrfail) {
 	TRACE(("enter send_msg_userauth_failure"))
 
 	CHECKCLEARTOWRITE();
-	
+
 	buf_putbyte(ses.writepayload, SSH_MSG_USERAUTH_FAILURE);
 
 	/* put a list of allowed types */
@@ -369,7 +369,7 @@ void send_msg_userauth_failure(int partial, int incrfail) {
 			buf_putbyte(typebuf, ',');
 		}
 	}
-	
+
 	if (ses.authstate.authtypes & AUTH_TYPE_PASSWORD) {
 		buf_putbytes(typebuf, (const unsigned char *)AUTH_METHOD_PASSWORD, AUTH_METHOD_PASSWORD_LEN);
 	}
@@ -410,7 +410,7 @@ void send_msg_userauth_failure(int partial, int incrfail) {
 		genrandom((unsigned char*)&rand_delay, sizeof(rand_delay));
 		rand_delay = mindelay + (rand_delay % vardelay);
 
-		if (delay.tv_sec == 0 && delay.tv_nsec <= mindelay) {
+		if (delay.tv_sec == 0 && delay.tv_nsec <= (int)mindelay) {
 			/* Compensate for elapsed time */
 			delay.tv_nsec = rand_delay - delay.tv_nsec;
 		} else {
@@ -443,7 +443,7 @@ void send_msg_userauth_failure(int partial, int incrfail) {
 		dropbear_exit("Max auth tries reached - user '%s' from %s",
 				userstr, svr_ses.addrstring);
 	}
-	
+
 	TRACE(("leave send_msg_userauth_failure"))
 }
 
@@ -457,7 +457,7 @@ void send_msg_userauth_success() {
 	buf_putbyte(ses.writepayload, SSH_MSG_USERAUTH_SUCCESS);
 	encrypt_packet();
 
-	/* authdone must be set after encrypt_packet() for 
+	/* authdone must be set after encrypt_packet() for
 	 * delayed-zlib mode */
 	ses.authstate.authdone = 1;
 	ses.connect_time = 0;
@@ -469,7 +469,7 @@ void send_msg_userauth_success() {
 
 	/* Remove from the list of pre-auth sockets. Should be m_close(), since if
 	 * we fail, we might end up leaking connection slots, and disallow new
-	 * logins - a nasty situation. */							
+	 * logins - a nasty situation. */
 	m_close(svr_ses.childpipe);
 
 	TRACE(("leave send_msg_userauth_success"))
